+++ packages/filterpane/src/hooks/useFilterContext.tsx
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { default as filterContext, FilterContext } from '../FilterContext';
+
+const useFilterContext = (): FilterContext => useContext(filterContext);
+
+export default useFilterContext;