+++ packages/filterpane/src/types/FilterOption.ts
@@ -1,0 +1,16 @@
+import { ReactNode } from 'react';
+
+type FilterOption = {
+  key: string;
+  label: ReactNode;
+  searchString?: string;
+};
+
+export type FilterOptions = {
+  options: Record<string, FilterOption>;
+  sortOrder?: string[];
+};
+
+export type FilterOptionsFn<TData> = (data: TData) => FilterOptions;
+
+export default FilterOption;