+++ packages/table/src/types.ts
@@ -1,3 +1,5 @@
+/* eslint-disable @typescript-eslint/ban-ts-comment */
+
 import {
   Column as ColumnDefault,
   PluginHook as PluginHookDefault,
@@ -77,4 +79,4 @@ declare module 'react-table' {
   }
 }
 
-export { TableOptions, CellProps } from 'react-table';
+export { TableOptions, CellProps, SortByFn } from 'react-table';