+++ storybook/src/components/ChangeLog/index.ts
@@ -1,0 +1,1 @@
+export { ChangeLog, ChangeLog as default } from './ChangeLog';