+++ storybook/src/stories/checkbox/Checkbox.template.tsx
@@ -1,0 +1,5 @@
+import { Checkbox, CheckboxProps } from '@equinor/fusion-react-checkbox';
+
+export const CheckboxTemplate = (props: CheckboxProps): React.ReactElement => <Checkbox {...props} />;
+
+export default CheckboxTemplate;