+++ storybook/stories/checkbox/Checkbox.template.tsx
@@ -1,5 +1,0 @@
-import { Checkbox, CheckboxProps } from '@equinor/fusion-react-checkbox';
-
-export const CheckboxTemplate = (props: CheckboxProps): JSX.Element => <Checkbox {...props} />;
-
-export default CheckboxTemplate;