+++ packages/date/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.4](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-date@0.3.3...@equinor/fusion-react-date@0.3.4) (2021-09-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-date
+
+
+
+
+
+## 0.3.3 (2021-09-16)
+
+**Note:** Version bump only for package @equinor/fusion-react-date
+
+
+
+
+
+## 0.3.2 (2021-09-10)
+
+**Note:** Version bump only for package @equinor/fusion-react-date
+
+
+
+
+
 ## 0.3.1 (2021-09-07)
 
 **Note:** Version bump only for package @equinor/fusion-react-date