+++ packages/datepicker/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.9.5](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-datepicker@0.9.4...@equinor/fusion-react-datepicker@0.9.5) (2021-09-23)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
+## 0.9.4 (2021-09-16)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
+## 0.9.3 (2021-09-08)
+
+**Note:** Version bump only for package @equinor/fusion-react-datepicker
+
+
+
+
+
 ## [0.9.2](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-datepicker@0.9.1...@equinor/fusion-react-datepicker@0.9.2) (2021-09-07)