+++ packages/textarea/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [0.4.0](https://github.com/equinor/fusion-react-components/compare/@equinor/fusion-react-textarea@0.3.2...@equinor/fusion-react-textarea@0.4.0) (2021-09-23)
+
+
+### Features
+
+* **textarea:** expose validation ([99a8842](https://github.com/equinor/fusion-react-components/commit/99a88423d79829f7e499d3d8a3f5ec1be105adb3))
+
+
+
+
+
+## 0.3.2 (2021-09-16)
+
+**Note:** Version bump only for package @equinor/fusion-react-textarea
+
+
+
+
+
 ## 0.3.1 (2021-09-07)
 
 **Note:** Version bump only for package @equinor/fusion-react-textarea